Refactor codebase
The codebase is currently split into 2 packages: chungus and
gui.

This makes nagivation and maintenance harder as the chungus
package is so large it blurs boundaries.

As a step towards better code quality, let's refactor the
chungus package into several smaller packages like app, logic,
model, common etc.

This allows for separation of concerns and an improvement in
readability.
