+++ src/main/java/chungus/model/Event.java
@@ -1,4 +1,4 @@
-package chungus;
+package chungus.model;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
@@ -82,3 +82,5 @@ public class Event extends Task {
                 toDisplay);
     }
 }
+
+
