+++ src/test/java/chungus/ParserTest.java
@@ -5,6 +5,9 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import org.junit.jupiter.api.Test;
 
+import chungus.common.ChungusException;
+import chungus.logic.Parser;
+
 public class ParserTest {
 
     @Test
